+++ src/onnx/runtime.rs
@@ -153,6 +153,7 @@ pub trait InferenceRuntime: Send {
 }
 
 #[cfg(test)]
+#[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;